fix(repo): m4 slice 5, security review findings fixed (#32)
Two independent adversarial reviews (guest side and host side, fresh contexts,
instructed to refute the safety claims) ran against the full tree. Both came
back NEEDS-CHANGES. Every blocking item is fixed here, with regression tests,
and docs/security-review-m4.md records the findings, dispositions, and the
invariants both reviews confirmed against the code.

The blocker: a hostile 64-bit CBOR length argument crashed the helper before
the token gate. Wire.token fully decodes the map ahead of the AuthGate, and
Int(UInt64) traps in Swift, so a tokenless ~12-byte frame killed the process.
Fixed by bounding every length and count argument against the remaining input
before any Int conversion, and by making CBORMap.int throw on integers outside
Int64 instead of trapping. The C codec had the mirror bug, an additive bounds
check that wraps (r->off + va), fixed in subtraction form in both readers.
Regression tests on both codecs craft the wrapping lengths and demand a clean
error.

The majors: accepted connections had no read deadline, so a connect-then-stall
flood could park unbounded serve threads (SO_RCVTIMEO and SO_SNDTIMEO now set
on every accepted fd); and the CLI helper never removed its token file, so the
session credential outlived the session and O_EXCL refused the next start
(SIGINT/SIGTERM handlers plus atexit now remove it; a SIGKILL still leaves the
file and the next start refuses loudly, keeping the M1 never-truncate call).

Also folded in: lifecycle state in LoopbackListener now mutates under the lock
with a double-start guard; the shadow registry says so on stderr when its
table is full; the message-algorithm hash refuses a >4 GiB input that would
truncate CC_LONG; the frame length assembles in uint32_t; and the SecItem
predicates document the tag-uniqueness assumption.

Verification: ctest 3/3 with the new wrapping-length test, protocol swift 28/0
including the four new hostile-input tests, helper and host-core build. The
SEP-backed suites stay blocked by the locked login session; full re-run queued
for the M4 closeout.

+++ packages/protocol/c/src/se_framing.c
@@ -30,8 +30,10 @@ int se_frame(const uint8_t *payload, size_t len, uint8_t *out, size_t cap) {
 }
 
 long se_payload_length(const uint8_t prefix[4]) {
-  long len =
-      ((long)prefix[0] << 24) | ((long)prefix[1] << 16) | ((long)prefix[2] << 8) | (long)prefix[3];
+  // Unsigned assembly: a signed shift of prefix[0] >= 0x80 would be negative on
+  // a 32-bit long. Apple targets are 64-bit, but the codec stays width-clean.
+  uint32_t len = ((uint32_t)prefix[0] << 24) | ((uint32_t)prefix[1] << 16) |
+                 ((uint32_t)prefix[2] << 8) | (uint32_t)prefix[3];
   if (len > SE_MAX_FRAME) return -1;
-  return len;
+  return (long)len;
 }
